+++ .eslintignore
@@ -4,6 +4,7 @@
 examples/cactus-example-carbon-accounting-frontend/www/
 examples/cactus-example-supply-chain-frontend/www/
 weaver/core/identity-management/iin-agent/out/
+weaver/common/protos-js/build/
 
 **/src/main/typescript/generated/proto/**
 **/src/main/typescript/generated/wasm-pack/**
